feat: Update to run with Horizon 32+ as unprivileged user
Upgrade the installer script for Ubuntu and Debian to run as an unprivileged user with sudo. Improved input to set
postgres root and OpenNMS database credentials which require confirmation to avoid mistakes.
OpenJDK 17 JDK is now installed automatically.

OpenNMS data base credentials will be stored in a secure vault instead of plain text in the opennms-datasources.xml.

Tested with Ubuntu 22.04 LTS cloud image